reply_to_mention/requirements.txt
anthropic==0.37.1 
python-dotenv==1.0.0
pyyaml==6.0.1
requests-oauthlib==2.0.0
supabase==2.10.0